backend/requirements.txt
anthropic==0.94.1
fastapi==0.135.2
uvicorn==0.42.0
requests==2.32.4
beautifulsoup4==4.13.4
python-multipart==0.0.22
httpx==0.28.1
pillow==10.4.0
python-dotenv==1.2.2
pydantic==2.12.5